[refactor] Int Extension을 Extension 폴더로 위치 변경
